Getting samples folder dynamically
